+++ .buildkite/serve.rayci.yml
@@ -3,7 +3,7 @@ steps:
   - label: ":ray-serve: serve: core tests"
     instance_type: medium
     commands:
-      - bazel run //ci/ray_ci -- //python/ray/tests/... serve --except-tags worker-container
+      - bazel run //ci/ray_ci:test_in_docker -- //python/ray/tests/... serve --except-tags worker-container
     depends_on: servebuild
     job_env: servebuild
 
@@ -11,7 +11,7 @@ steps:
     parallelism: 3
     instance_type: large
     commands:
-      - bazel run //ci/ray_ci -- //python/ray/serve/... serve --except-tags post_wheel_build,gpu,xcommit
+      - bazel run //ci/ray_ci:test_in_docker -- //python/ray/serve/... serve --except-tags post_wheel_build,gpu,xcommit
         --workers "$${BUILDKITE_PARALLEL_JOB_COUNT}" --worker-id "$${BUILDKITE_PARALLEL_JOB}" --parallelism-per-worker 3
     depends_on: servebuild
     job_env: servebuild
